requirements.txt
Flask
Pillow
gunicorn
PyMuPDF